requirements.txt
boto3
connexion
gevent
requests
stups-tokens
